gain uint32_t

+++ spi/spi_master_slave/spi_master/loop/main_core.cpp
@@ -294,7 +294,7 @@ case SENDDATALIN:
 case SET_PID_GAIN:
               {
                 ALGCODE=ALGNONE;
-                if (HARDWAREVERSION!=BBFPGA) scanner->hardware->set_GainPID((uint16_t)vector[1]);
+                if (HARDWAREVERSION!=BBFPGA) scanner->hardware->set_GainPID((uint32_t)vector[1]);
                 else                         scanner->hardware->set_GainPID((uint32_t)vector[1]); 
                 break; 
               }
